refactor: extract skill logic to module
